Move devlog into debug mode so it is not included at build time for security & to avoid footguns.
