+++ toolruntime/debug_disabled.go
@@ -2,6 +2,8 @@
 
 package toolruntime
 
+import "time"
+
 // debugEnabled is a compile-time false constant in production builds. Every
 // `if debugEnabled { ... }` block and every call whose body is purely
 // `if !debugEnabled { return ... }` collapses at compile time, so the
@@ -17,3 +19,26 @@ func debugLogf(string, ...any) {}
 func debugPreview(any, int) string { return "" }
 
 func debugMessagesSummary([]any, int) string { return "" }
+
+// ---------------------------------------------------------------------------
+// devLog — no-op stub for production builds
+// ---------------------------------------------------------------------------
+
+// devLog is a no-op stub compiled into production binaries (no
+// toolruntime_debug build tag). Every method is a no-op so call sites
+// compile unconditionally without any #ifdef ceremony. The real
+// implementation lives in debug_enabled.go and is only included in
+// debug builds, so user content never touches disk in a deployed
+// enclave regardless of the runtime DEBUG flag.
+type devLog struct{}
+
+func openDevLog(string) *devLog                                                       { return nil }
+func (d *devLog) Close()                                                              {}
+func (d *devLog) WriteHeader(map[string]any, string, string, string)                  {}
+func (d *devLog) WriteTurnHeader(int)                                                 {}
+func (d *devLog) WriteTokens(map[string]any)                                          {}
+func (d *devLog) WriteResponseBody(map[string]any)                                    {}
+func (d *devLog) WriteStreamedThinkingAndContent(string, string)                      {}
+func (d *devLog) WriteToolCalls([]toolCall)                                           {}
+func (d *devLog) WriteToolExec(string, map[string]any, string, time.Duration, string) {}
+func (d *devLog) WriteFinish(string)                                                  {}
